+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 ### :boom: Breaking changes & Deprecations
 ### :sparkles: New features
+- Better overlays interactions on boards inside boards [Taiga #4386](https://tree.taiga.io/project/penpot/us/4386)
 ### :bug: Bugs fixed
 
 - Add title to color bullets [Taiga #4218](https://tree.taiga.io/project/penpot/task/4218)